Extract ethereum-related code to separate packages
