+++ test/builders/blocks.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"github.com/orbs-network/crypto-lib-go/crypto/digest"
 	"github.com/orbs-network/crypto-lib-go/crypto/hash"
-	"github.com/orbs-network/crypto-lib-go/test/crypto/keys"
+	"github.com/orbs-network/crypto-lib-go/test/crypto/ethereum/keys"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
 	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/orbs-network/orbs-spec/types/go/services"
